requirements.txt
black==24.4.2
Django==4.2
ipython==8.12.3
pillow==10.4.0
psycopg2==2.9.9
redis==5.0.8
python-dotenv==1.0.1
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
isort==5.13.2
